Rename files to use UpperCamelCaseCluster/Logic for descriptor
