+++ src/app/clusters/descriptor/app_config_dependent_sources.cmake
@@ -23,6 +23,6 @@ TARGET_SOURCES(
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
-    "${CLUSTER_DIR}/descriptor-cluster.h"
-    "${CLUSTER_DIR}/descriptor-cluster.cpp"
+    "${CLUSTER_DIR}/DescriptorCluster.h"
+    "${CLUSTER_DIR}/DescriptorCluster.cpp"
 )
